requirements.txt
langchain_openai 
langchain_core
langchain_community
python-dotenv
streamlit
langserve
fastapi
bs4
pypdf
chromadb
faiss-cpu
beautifulsoup4
langserve